add facility for copying lib version to xll version
You can now mark ExternalLibrary elements in the .dna file with UseVersionAsOutputVersion="true" to copy their version resource into the output .xll. If you specify this for more than one ExternalLibrary only the first will be used, and a warning will be emitted. If the version copying fails for some reason (like the source file doesn't have it, e.g. if it's a .dna file), then a warning is emitted.